frontend/cli/pyproject.toml
[project]
name = "gwop"
version = "1.0.1"
readme = "README.md"
license = {text = "MIT"}
description = "gwop CLI"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]


[project.scripts]
gwop = "gwop.gwop:gwop"

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
